MMT-2798 - fixed tests
